mental_health_system/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
transformers>=4.36,<5.0
torch>=2.1,<3.0
numpy>=1.26,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
